added clients.js with GQClient and MockClient
