+++ src/lib/bind.js
@@ -1,28 +1,11 @@
 import { writable } from 'svelte/store';
 import { buffer, switchMap, from, interval, NEVER, Subject, debounceTime, skip } from 'rxjs';
 import { onDestroy } from 'svelte';
-import { GraphQLClient } from 'graphql-request'
 import {error, success} from '$lib/store';
 
 const T = 2000;
 
-export function GQClient({apiServerUrl, token, put, post, setId}){
-  return async function(){
-    if(token){
-      const t = await token();
-      let c = new GraphQLClient(apiServerUrl, { headers: {Authorization: `Bearer ${t}`} })
-    }else{
-      let c = new GraphQLClient(apiServerUrl)
-    }
-    return {
-        put: (values, id) => put(values, id, c),
-        post: (values, c) => post(values, c),
-        setId
-    }
-  }
-}
-
-export default function({id, client}){
+export function stream({id, client}){
 
     const status = writable('initial')
