+++ src/lib/index.js
@@ -1,0 +1,4 @@
+import { stream } from './bind'
+import { GQClient, MockClient } from './clients'
+
+export {GQClient, MockClient, stream}
